Some refactoring and re-arranging. Extra cleanup from large vessel - moved trait time re-calculation to helper method. Inlined large vessel special capability class. Fixed infinite decay tooltip from displaying weird values at Long.MAX_VALUE. Fixed config gui not displaying due to incorrect annotation
