+++ src/main/java/net/dries007/tfc/client/TFCSoundEvents.java
@@ -3,7 +3,7 @@
  * See the project README.md and LICENSE.txt for more information.
  */
 
-package net.dries007.tfc.util;
+package net.dries007.tfc.client;
 
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.SoundEvent;
