+++ src/main/java/net/dries007/tfc/objects/LootTablesTFC.java
@@ -3,14 +3,14 @@
  * See the project README.md and LICENSE.txt for more information.
  */
 
-package net.dries007.tfc.util;
+package net.dries007.tfc.objects;
 
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.storage.loot.LootTableList;
 
 import static net.dries007.tfc.api.util.TFCConstants.MOD_ID;
 
-public class LootTableListTFC
+public class LootTablesTFC
 {
     public static final ResourceLocation ANIMALS_BEAR = register("animals/bear");
     public static final ResourceLocation ANIMALS_CHICKEN = register("animals/chicken");
