+++ src/main/java/net/dries007/tfc/objects/Powder.java
@@ -18,6 +18,5 @@ public enum Powder
     LIMONITE_POWDER,
     MALACHITE_POWDER,
     SALT,
-    FERTILIZER,
-    ;
+    FERTILIZER
 }
