+++ src/main/java/net/dries007/tfc/objects/entity/animal/EntityBearTFC.java
@@ -26,8 +26,8 @@ import net.minecraft.world.World;
 import net.minecraft.world.biome.Biome;
 
 import net.dries007.tfc.Constants;
-import net.dries007.tfc.util.LootTableListTFC;
-import net.dries007.tfc.util.TFCSoundEvents;
+import net.dries007.tfc.client.TFCSoundEvents;
+import net.dries007.tfc.objects.LootTablesTFC;
 import net.dries007.tfc.util.calendar.CalendarTFC;
 import net.dries007.tfc.world.classic.biomes.BiomesTFC;
 
@@ -154,7 +154,7 @@ public class EntityBearTFC extends EntityAnimalMammal implements IMob, IAnimalTF
     @Nullable
     protected ResourceLocation getLootTable()
     {
-        return LootTableListTFC.ANIMALS_BEAR;
+        return LootTablesTFC.ANIMALS_BEAR;
     }
 
     @Override
