+++ src/main/java/net/dries007/tfc/objects/entity/animal/EntityDeerTFC.java
@@ -22,9 +22,9 @@ import net.minecraft.world.World;
 import net.minecraft.world.biome.Biome;
 
 import net.dries007.tfc.Constants;
+import net.dries007.tfc.client.TFCSoundEvents;
+import net.dries007.tfc.objects.LootTablesTFC;
 import net.dries007.tfc.objects.items.ItemsTFC;
-import net.dries007.tfc.util.LootTableListTFC;
-import net.dries007.tfc.util.TFCSoundEvents;
 import net.dries007.tfc.util.calendar.CalendarTFC;
 
 @ParametersAreNonnullByDefault
@@ -140,7 +140,7 @@ public class EntityDeerTFC extends EntityAnimalMammal implements IAnimalTFC
     @Nullable
     protected ResourceLocation getLootTable()
     {
-        return LootTableListTFC.ANIMALS_DEER;
+        return LootTablesTFC.ANIMALS_DEER;
     }
 
     @Override
