+++ src/main/java/net/dries007/tfc/objects/entity/animal/EntitySheepTFC.java
@@ -34,8 +34,8 @@ import net.minecraftforge.common.IShearable;
 import net.minecraftforge.oredict.OreDictionary;
 
 import net.dries007.tfc.Constants;
+import net.dries007.tfc.objects.LootTablesTFC;
 import net.dries007.tfc.objects.items.ItemsTFC;
-import net.dries007.tfc.util.LootTableListTFC;
 import net.dries007.tfc.util.calendar.CalendarTFC;
 import net.dries007.tfc.world.classic.biomes.BiomesTFC;
 
@@ -243,7 +243,7 @@ public class EntitySheepTFC extends EntityAnimalMammal implements IShearable, IA
     @Nullable
     protected ResourceLocation getLootTable()
     {
-        return LootTableListTFC.ANIMALS_SHEEP;
+        return LootTablesTFC.ANIMALS_SHEEP;
     }
 
     @Override
